Move CaptureP from factfilter.dsl package to the core.algebra
