+++ core/src/main/scala/com/rallyhealth/vapors/factfilter/dsl/ExprBuilder.scala
@@ -1,7 +1,7 @@
 package com.rallyhealth.vapors.factfilter.dsl
 
 import cats.{FlatMap, Foldable, Functor, FunctorFilter, Id, Order, Traverse, TraverseFilter}
-import com.rallyhealth.vapors.core.algebra.Expr
+import com.rallyhealth.vapors.core.algebra.{CaptureP, Expr}
 import com.rallyhealth.vapors.core.data.{Evidence, TypedFact, Window}
 import com.rallyhealth.vapors.core.lens.NamedLens
 import com.rallyhealth.vapors.core.math.{Addition, Negative, Subtraction}
